Added reletaive path for log files, added conditional delete to merge pdfs, updated gitignore, removed pycache files, updated create directory util, added create directory to merge file, added flag for delete files to merge file

+++ pdf_utils/merge_pdfs.py
@@ -1,13 +1,15 @@
 import os
 import PyPDF2
 import logging.config
-from royal_road.config import PDF_OUTPUT_DIR, LOGGING_CONFIG
+from royal_road.config import PDF_OUTPUT_DIR, LOGGING_CONFIG, MERGED_PDF_OUTPUT_DIR
+from royal_road.utils.utils import create_directory
 
 # If you want to use a different log  file
 # LOGGING_CONFIG['handlers']['file']['filename'] = '../logs/scrape.log'
 
 # Configure logging using the dictionary
 logging.config.dictConfig(LOGGING_CONFIG)
+
 
 def merge_pdfs(directory_path, output_path):
     merger = PyPDF2.PdfMerger()
@@ -28,6 +30,7 @@ def merge_pdfs(directory_path, output_path):
 
     return output_path
 
+
 def delete_source_files(directory_path, exclude_file):
     # Get all PDF files in the directory
     pdf_files = [f for f in os.listdir(directory_path) if f.endswith('.pdf')]
@@ -39,22 +42,31 @@ def delete_source_files(directory_path, exclude_file):
             os.remove(file_path)
             logging.info(f"Deleted file: {filename}")
 
-# Adjust paths and filenames based on configuration
-pdf_directory = PDF_OUTPUT_DIR
-fiction_directory = 'He_Who_Fights_With_Monsters'
-output_filename = '000_merged.pdf'
-input_path = os.path.join(pdf_directory, fiction_directory)
-output_full_path = os.path.join(input_path, output_filename)
 
-# Merge PDFs
-try:
-    merged_pdf_path = merge_pdfs(input_path, output_full_path)
-    logging.info(f"***************Starting Merge***************")
-    logging.info(f'PDFs merged successfully to: {merged_pdf_path}')
+def main(delete_files: bool = False):
+    # Adjust paths and filenames based on configuration
+    pdf_directory = PDF_OUTPUT_DIR
+    merge_directory = MERGED_PDF_OUTPUT_DIR
+    fiction_directory = 'He_Who_Fights_With_Monsters'
+    output_filename = 'BookNameHere.pdf'
+    input_path = os.path.join(pdf_directory, fiction_directory)
+    output_directory = os.path.join(merge_directory, fiction_directory)
+    output_full_path = os.path.join(merge_directory, fiction_directory, output_filename)
 
-    # Delete source PDF files excluding the output file
-    logging.info(f"***************Starting Purge***************")
-    delete_source_files(input_path, output_full_path)
-    logging.info('Source files deleted successfully.')
-except Exception as e:
-    logging.error(f"Error occurred during merging or deleting files: {str(e)}")
+    # Merge PDFs
+    try:
+        create_directory(output_directory)
+        merged_pdf_path = merge_pdfs(input_path, output_full_path)
+        logging.info(f"***************Starting Merge***************")
+        logging.info(f'PDFs merged successfully to: {merged_pdf_path}')
+
+        # Delete source PDF files excluding the output file
+        if delete_files:
+            logging.info(f"***************Starting Purge***************")
+            delete_source_files(input_path, output_full_path)
+            logging.info(f"***************Ending Purge***************")
+    except Exception as e:
+        logging.error(f"Error occurred during merging or deleting files: {str(e)}")
+
+if __name__ == '__main__':
+    main(delete_files=False)
